Updating gRPC code to use the latest Netty with application-level flow control support.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=80230204

+++ netty/src/test/java/com/google/net/stubby/transport/netty/NettyClientHandlerTest.java
@@ -134,7 +134,7 @@ public class NettyClientHandlerTest extends NettyHandlerTestBase {
 
     handler.write(ctx, new CancelStreamCommand(stream), promise);
 
-    ByteBuf expected = rstStreamFrame(3, Http2Error.CANCEL.code());
+    ByteBuf expected = rstStreamFrame(3, (int) Http2Error.CANCEL.code());
     verify(ctx).write(eq(expected), eq(promise));
     verify(ctx).flush();
   }
